requirements.txt
pipecat-ai-flows
pipecat-ai[daily,silero,azure,google,openai,aic,koala,anthropic,cartesia,elevenlabs,deepgram,local-smart-turn-v3,runner,webrtc,simli]==0.0.100
python-dotenv>=1.0.0
loguru>=0.7.0
aiohttp>=3.9.0
supabase>=2.3.0
openai>=1.12.0
azure-cognitiveservices-speech>=1.35.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
requests>=2.31.0